micam: t4ka3: (match-upstream) remove non-gmin codes
Removing non-gmin codes fixes the following build error:

        micam/atomisp-t4ka3.c:507:27: error: ‘struct camera_sensor_platform_data’ has no member named ‘power_ctrl’
          507 |         dev->platform_data->power_ctrl(sd, 0);
              |

The upstream atomisp is not intended to be used with non-gmin platform
codes. Just remove them.

By the way, this seems to be a real bug on Xiaomi Mi Pad 2. It uses
gmin, so there is no member named power_ctrl() in platform_data I guess.
But the code is not guarded for gmin.

Signed-off-by: Tsuchiya Yuto <kitakar@gmail.com>